aplit backend and attester api | #1694
