+++ extension/chrome/settings/modules/my_key.ts
@@ -8,6 +8,8 @@ import { Att } from '../../../js/common/core/att.js';
 import { Ui, Env, Browser } from '../../../js/common/browser.js';
 import { Pgp } from '../../../js/common/core/pgp.js';
 import { Api } from '../../../js/common/api/api.js';
+import { Attester } from '../../../js/common/api/attester.js';
+import { Backend } from '../../../js/common/api/backend.js';
 
 declare const openpgp: typeof OpenPGP;
 declare const ClipboardJS: any;
@@ -29,8 +31,8 @@ Catch.try(async () => {
   const { keys: [prv] } = await openpgp.key.readArmored(primaryKi.private);
 
   try {
-    const result = await Api.attester.lookupEmail(acctEmail);
-    const url = Api.fc.url('pubkey', acctEmail);
+    const result = await Attester.attester.lookupEmail(acctEmail);
+    const url = Backend.fc.url('pubkey', acctEmail);
     if (result.pubkey && await Pgp.key.longid(result.pubkey) === primaryKi.longid) {
       $('.pubkey_link_container a').text(url.replace('https://', '')).attr('href', url).parent().css('visibility', 'visible');
     }
