+++ extension/js/background_page/migrations.ts
@@ -7,6 +7,8 @@ import { Store } from '../common/platform/store.js';
 import { Value, Str } from '../common/core/common.js';
 import { Api } from '../common/api/api.js';
 import { Pgp } from '../common/core/pgp.js';
+import { Attester } from '../common/api/attester.js';
+import { Backend } from '../common/api/backend.js';
 
 export const migrateGlobal = async () => {
   if (window.localStorage && window.localStorage.length > 0) { // window.localStorage may be null on Firefox, likely disabled in settings?
@@ -42,7 +44,7 @@ const accountUpdateStatusKeyserver = async (acctEmail: string) => { // checks wh
       await Store.setAcct(acctEmail, storage); // fix duplicate email addresses
     }
     try {
-      const lookupEmailsRes = await Api.attester.lookupEmails(storage.addresses);
+      const lookupEmailsRes = await Attester.attester.lookupEmails(storage.addresses);
       const addressesKeyserver = [];
       for (const email of Object.keys(lookupEmailsRes)) {
         const result = lookupEmailsRes[email];
@@ -67,11 +69,11 @@ export const scheduleFcSubscriptionLevelCheck = (bgProcessStartReason: 'update' 
   Catch.setHandledTimeout(() => {
     if (bgProcessStartReason === 'update' || bgProcessStartReason === 'chrome_update') {
       // update may happen to too many people at the same time -- server overload
-      Catch.setHandledTimeout(() => Api.fc.accountCheckSync().catch(reportSignificantErrs), Value.int.hoursAsMiliseconds(Math.random() * 3)); // random 0-3 hours
+      Catch.setHandledTimeout(() => Backend.fc.accountCheckSync().catch(reportSignificantErrs), Value.int.hoursAsMiliseconds(Math.random() * 3)); // random 0-3 hours
     } else {
       // the user just installed the plugin or started their browser, no risk of overloading servers
-      Api.fc.accountCheckSync().catch(reportSignificantErrs); // now
+      Backend.fc.accountCheckSync().catch(reportSignificantErrs); // now
     }
   }, 10 * 60 * 1000); // 10 minutes
-  Catch.setHandledInterval(() => Api.fc.accountCheckSync().catch(reportSignificantErrs), Value.int.hoursAsMiliseconds(23 + Math.random())); // random 23-24 hours
+  Catch.setHandledInterval(() => Backend.fc.accountCheckSync().catch(reportSignificantErrs), Value.int.hoursAsMiliseconds(23 + Math.random())); // random 23-24 hours
 };
