+++ extension/js/common/account.ts
@@ -8,6 +8,7 @@ import { Api } from './api/api.js';
 import { Env } from './browser.js';
 import { Google } from './api/google.js';
 import { Buf } from './core/buf.js';
+import { Backend } from './api/backend.js';
 
 type AccountEventHandlersOptional = {
   renderStatusText?: (text: string, showSpinner?: boolean) => void;
@@ -47,7 +48,7 @@ export class FcAcct {
 
   subscribe = async (acctEmail: string, chosenProduct: Product, source: string | undefined) => {
     this.eventHandlers.renderStatusText(chosenProduct.method === 'trial' ? 'enabling trial..' : 'upgrading..', true);
-    await Api.fc.accountCheckSync();
+    await Backend.fc.accountCheckSync();
     try {
       const newSubscriptionInfo = await this.doSubscribe(chosenProduct, source);
       const globalStoreUpdate: GlobalStore = {};
@@ -68,7 +69,7 @@ export class FcAcct {
 
   register = async (acctEmail: string) => { // register_and_attempt_to_verify
     this.eventHandlers.renderStatusText('registering..', true);
-    const response = await Api.fc.accountLogin(acctEmail);
+    const response = await Backend.fc.accountLogin(acctEmail);
     if (response.verified) {
       return response;
     }
@@ -90,7 +91,7 @@ export class FcAcct {
     let lastTokenErr;
     for (const token of tokens) {
       try {
-        return await Api.fc.accountLogin(acctEmail, token);
+        return await Backend.fc.accountLogin(acctEmail, token);
       } catch (e) {
         if (Api.err.isStandardErr(e, 'token')) {
           lastTokenErr = e;
@@ -127,7 +128,7 @@ export class FcAcct {
   private doSubscribe = async (chosenProduct: Product, source?: string) => {
     await Store.removeGlobal(['cryptup_subscription_attempt']);
     // todo - deal with auth error? would need to know account_email for new registration
-    const response = await Api.fc.accountSubscribe(chosenProduct.id!, chosenProduct.method!, source);
+    const response = await Backend.fc.accountSubscribe(chosenProduct.id!, chosenProduct.method!, source);
     if (response.subscription.level === chosenProduct.level && response.subscription.method === chosenProduct.method) {
       return response.subscription;
     }
