+++ extension/js/common/api/api.ts
@@ -4,23 +4,19 @@
 
 // tslint:disable:no-direct-ajax
 
-import { Store, GlobalStore, Serializable, Subscription } from '../platform/store.js';
+import { Store } from '../platform/store.js';
 import { Value, Str, Dict } from '../core/common.js';
-import { Pgp, Contact } from '../core/pgp.js';
-import { Ui, Xss, Env } from '../browser.js';
+import { Contact } from '../core/pgp.js';
+import { Xss, Env } from '../browser.js';
 import { Att } from '../core/att.js';
 import { SendableMsgBody } from '../core/mime.js';
-import { PaymentMethod } from '../account.js';
 import { Catch } from '../platform/catch.js';
 import { Buf } from '../core/buf.js';
 import { BrowserMsg } from '../extension.js';
 
 type StandardError = { code: number | null; message: string; internal: string | null; data?: string; stack?: string; };
 type StandardErrorRes = { error: StandardError };
-type FcAuthToken = { account: string, token: string };
-type FcAuthMethods = 'uuid' | FcAuthToken | null;
-type SubscriptionLevel = 'pro' | null;
-type ReqFmt = 'JSON' | 'FORM';
+export type ReqFmt = 'JSON' | 'FORM';
 type ResFmt = 'json';
 export type ReqMethod = 'POST' | 'GET' | 'DELETE' | 'PUT';
 export type ProviderContactsResults = { new: Contact[], all: Contact[] };
@@ -38,9 +34,6 @@ export type ProgressCb = (percent?: number, loaded?: number, total?: number) => 
 export type ProgressCbs = { upload?: ProgressCb | null, download?: ProgressCb | null };
 export type ProviderContactsQuery = { substring: string };
 export type SendableMsg = { headers: Dict<string>; from: string; to: string[]; subject: string; body: SendableMsgBody; atts: Att[]; thread?: string; };
-export type SubscriptionInfo = { active?: boolean | null; method?: PaymentMethod | null; level?: SubscriptionLevel; expire?: string | null; expired?: boolean };
-export type PubkeySearchResult = { pubkey: string | null; has_cryptup: boolean | null; };
-export type AwsS3UploadItem = { baseUrl: string, fields: { key: string; file?: Att }, att: Att };
 
 abstract class ApiCallError extends Error {
 
@@ -136,68 +129,6 @@ export class ApiErrorResponse extends ApiCallError {
 
 export class AuthError extends Error { }
 
-export namespace R { // responses
-
-  export type FcHelpFeedback = { sent: boolean };
-  export type FcAccountLogin = { registered: boolean, verified: boolean, subscription: SubscriptionInfo };
-  export type FcAccountUpdate$result = { alias: string, email: string, intro: string, name: string, photo: string, default_message_expire: number };
-  export type FcAccountUpdate = { result: FcAccountUpdate$result, updated: boolean };
-  export type FcAccountSubscribe = { subscription: SubscriptionInfo };
-  export type FcAccountCheck = { email: string | null, subscription: SubscriptionInfo | null };
-  export type FcBlogPost = { title: string, date: string, url: string };
-
-  export type FcMsgPresignFiles = { approvals: { base_url: string, fields: { key: string } }[] };
-  export type FcMsgConfirmFiles = { confirmed: string[], admin_codes: string[] };
-  export type FcMsgToken = { token: string };
-  export type FcMsgUpload = { short: string, admin_code: string };
-  export type FcLinkMsg = { expire: string, deleted: boolean, url: string, expired: boolean };
-  export type FcLinkMe$profile = {
-    alias: string | null, name: string | null, photo: string | null, intro: string | null, web: string | null,
-    phone: string | null, token: string | null, subscription_level: string | null, subscription_method: string | null, email: string | null
-  };
-  export type FcLinkMe = { profile: null | FcLinkMe$profile };
-  export type ApirFcMsgExpiration = { updated: boolean };
-
-  export type AttTestWelcome = { sent: boolean };
-  export type AttInitialLegacySugmit = { saved: boolean };
-  export type AttKeyserverDiagnosis = { hasPubkeyMissing: boolean, hasPubkeyMismatch: boolean, results: Dict<{ pubkey?: string, match: boolean }> };
-
-  export type GmailUsersMeProfile = { emailAddress: string, historyId: string, messagesTotal: number, threadsTotal: string };
-  export type GmailMsg$header = { name: string, value: string };
-  export type GmailMsg$payload$body = { attachmentId: string, size: number, data?: string };
-  export type GmailMsg$payload$part = { body?: GmailMsg$payload$body, filename?: string, mimeType?: string, headers?: GmailMsg$header[] };
-  export type GmailMsg$payload = { parts?: GmailMsg$payload$part[], headers?: GmailMsg$header[], mimeType?: string, body?: GmailMsg$payload$body };
-  export type GmailMsg$labelId = 'INBOX' | 'UNREAD' | 'CATEGORY_PERSONAL' | 'IMPORTANT' | 'SENT' | 'CATEGORY_UPDATES';
-  export type GmailMsg = {
-    id: string; historyId: string; threadId?: string | null; payload: GmailMsg$payload; internalDate?: number | string;
-    labelIds?: GmailMsg$labelId[]; snippet?: string; raw?: string;
-  };
-  export type GmailMsgList$message = { id: string, threadId: string };
-  export type GmailMsgList = { messages?: GmailMsgList$message[], resultSizeEstimate: number };
-  export type GmailLabels$label = {
-    id: string, name: string, messageListVisibility: 'show' | 'hide', labelListVisibility: 'labelShow' | 'labelHide', type: 'user' | 'system',
-    messagesTotal?: number, messagesUnread?: number, threadsTotal?: number, threadsUnread?: number, color?: { textColor: string, backgroundColor: string }
-  };
-  export type GmailLabels = { labels: GmailLabels$label[] };
-  export type GmailAtt = { attachmentId: string, size: number, data: Buf };
-  export type GmailMsgSend = { id: string };
-  export type GmailThread = { id: string, historyId: string, messages: GmailMsg[] };
-  export type GmailThreadList = { threads: { historyId: string, id: string, snippet: string }[], nextPageToken: string, resultSizeEstimate: number };
-  export type GmailDraftCreate = { id: string };
-  export type GmailDraftDelete = {};
-  export type GmailDraftUpdate = {};
-  export type GmailDraftGet = { id: string, message: GmailMsg };
-  export type GmailDraftSend = {};
-  export type GmailAliases = { sendAs: GmailAliases$sendAs[] };
-  type GmailAliases$sendAs = { sendAsEmail: string, displayName: string, replyToAddress: string, signature: string, isDefault: boolean, treatAsAlias: boolean, verificationStatus: string };
-
-  export type OpenId = { // 'name' is the full name, picture is url
-    at_hash: string; exp: number; iat: number; sub: string; aud: string; azp: string; iss: "https://accounts.google.com";
-    name: string; picture: string; locale: 'en' | string; family_name: string; given_name: string;
-  };
-
-}
-
 export class Api {
 
   public static err = {
@@ -267,7 +198,7 @@ export class Api {
       if (e instanceof ApiErrorResponse && typeof e.res === 'object' && typeof e.res.error === 'object' && e.res.error.internal === 'auth') {
         return true;
       }
-      if (Api.internal.isStandardError(e) && e.internal === internalType) {
+      if (Api.isStandardError(e) && e.internal === internalType) {
         return true;
       }
       if ((e as StandardErrorRes).error && typeof (e as StandardErrorRes).error === 'object' && (e as StandardErrorRes).error.internal === internalType) {
@@ -366,236 +297,6 @@ export class Api {
     },
   };
 
-  public static attester = {
-    lookupEmail: (email: string): Promise<PubkeySearchResult> => {
-      return Api.internal.apiAttesterCall('lookup/email', { email: Str.parseEmail(email).email });
-    },
-    lookupEmails: async (emails: string[]): Promise<Dict<PubkeySearchResult>> => {
-      const results: Dict<PubkeySearchResult> = {};
-      await Promise.all(emails.map(async (email: string) => {
-        results[email] = await Api.attester.lookupEmail(email);
-      }));
-      return results;
-    },
-    initialLegacySubmit: (email: string, pubkey: string): Promise<R.AttInitialLegacySugmit> => Api.internal.apiAttesterCall('initial/legacy_submit', {
-      email: Str.parseEmail(email).email,
-      pubkey: pubkey.trim(),
-      // attest: false,
-    }),
-    testWelcome: (email: string, pubkey: string): Promise<R.AttTestWelcome> => Api.internal.apiAttesterCall('test/welcome', {
-      email,
-      pubkey,
-    }),
-    diagnoseKeyserverPubkeys: async (acctEmail: string): Promise<R.AttKeyserverDiagnosis> => {
-      const diagnosis: R.AttKeyserverDiagnosis = { hasPubkeyMissing: false, hasPubkeyMismatch: false, results: {} };
-      const { addresses } = await Store.getAcct(acctEmail, ['addresses']);
-      const storedKeys = await Store.keysGet(acctEmail);
-      const storedKeysLongids = storedKeys.map(ki => ki.longid);
-      const results = await Api.attester.lookupEmails(Value.arr.unique([acctEmail].concat(addresses || [])));
-      for (const email of Object.keys(results)) {
-        const pubkeySearchResult = results[email];
-        if (!pubkeySearchResult.pubkey) {
-          diagnosis.hasPubkeyMissing = true;
-          diagnosis.results[email] = { pubkey: undefined, match: false };
-        } else {
-          let match = true;
-          if (!storedKeysLongids.includes(String(await Pgp.key.longid(pubkeySearchResult.pubkey)))) {
-            diagnosis.hasPubkeyMismatch = true;
-            match = false;
-          }
-          diagnosis.results[email] = { pubkey: pubkeySearchResult.pubkey, match };
-        }
-      }
-      return diagnosis;
-    },
-  };
-
-  public static fc = {
-    url: (type: string, variable = '') => {
-      return ({
-        'api': 'https://flowcrypt.com/api/',
-        'me': 'https://flowcrypt.com/me/' + variable,
-        'pubkey': 'https://flowcrypt.com/pub/' + variable,
-        'decrypt': 'https://flowcrypt.com/' + variable,
-        'web': 'https://flowcrypt.com/',
-      } as Dict<string>)[type];
-    },
-    helpFeedback: (acctEmail: string, message: string): Promise<R.FcHelpFeedback> => Api.internal.apiFcCall('help/feedback', {
-      email: acctEmail,
-      message,
-    }),
-    helpUninstall: (email: string, client: string): Promise<unknown> => Api.internal.apiFcCall('help/uninstall', {
-      email,
-      client,
-      metrics: null, // tslint:disable-line:no-null-keyword
-    }),
-    accountLogin: async (acctEmail: string, token?: string): Promise<{ verified: boolean, subscription: SubscriptionInfo }> => {
-      const authInfo = await Store.authInfo();
-      const uuid = authInfo.uuid || await Pgp.hash.sha1UtfStr(Pgp.password.random());
-      const account = authInfo.acctEmail || acctEmail;
-      const response = await Api.internal.apiFcCall('account/login', {
-        account,
-        uuid,
-        token: token || null, // tslint:disable-line:no-null-keyword
-      }) as R.FcAccountLogin;
-      if (response.registered !== true) {
-        throw new Error('account_login did not result in successful registration');
-      }
-      await Store.setGlobal({ cryptup_account_email: account, cryptup_account_uuid: uuid, cryptup_account_subscription: response.subscription });
-      return { verified: response.verified === true, subscription: response.subscription };
-    },
-    accountCheck: (emails: string[]) => Api.internal.apiFcCall('account/check', {
-      emails,
-    }) as Promise<R.FcAccountCheck>,
-    accountCheckSync: async () => { // callbacks true on updated, false not updated, null for could not fetch
-      const emails = await Store.acctEmailsGet();
-      if (emails.length) {
-        const response = await Api.fc.accountCheck(emails);
-        const authInfo = await Store.authInfo();
-        const subscription = await Store.subscription();
-        const globalStoreUpdate: GlobalStore = {};
-        if (response.email) {
-          if (response.email !== authInfo.acctEmail) {
-            // will fail auth when used on server, user will be prompted to verify this new device when that happens
-            globalStoreUpdate.cryptup_account_email = response.email;
-            globalStoreUpdate.cryptup_account_uuid = await Pgp.hash.sha1UtfStr(Pgp.password.random());
-          }
-        } else {
-          if (authInfo.acctEmail) {
-            globalStoreUpdate.cryptup_account_email = undefined;
-            globalStoreUpdate.cryptup_account_uuid = undefined;
-          }
-        }
-        Subscription.updateSubscriptionGlobalStore(globalStoreUpdate, subscription, response.subscription);
-        if (Object.keys(globalStoreUpdate).length) {
-          Catch.log('updating account subscription from ' + subscription.level + ' to ' + (response.subscription ? response.subscription.level : undefined), response);
-          await Store.setGlobal(globalStoreUpdate);
-          return true;
-        } else {
-          return false;
-        }
-      }
-      return undefined;
-    },
-    accountUpdate: async (updateValues?: Dict<Serializable>): Promise<R.FcAccountUpdate> => {
-      const authInfo = await Store.authInfo();
-      const request = { account: authInfo.acctEmail, uuid: authInfo.uuid } as Dict<Serializable>;
-      if (updateValues) {
-        for (const k of Object.keys(updateValues)) {
-          request[k] = updateValues[k];
-        }
-      }
-      return await Api.internal.apiFcCall('account/update', request) as R.FcAccountUpdate;
-    },
-    accountSubscribe: async (product: string, method: string, paymentSourceToken?: string): Promise<R.FcAccountSubscribe> => {
-      const authInfo = await Store.authInfo();
-      const response = await Api.internal.apiFcCall('account/subscribe', {
-        account: authInfo.acctEmail,
-        uuid: authInfo.uuid,
-        method,
-        source: paymentSourceToken || null, // tslint:disable-line:no-null-keyword
-        product,
-      }) as R.FcAccountSubscribe;
-      await Store.setGlobal({ cryptup_account_subscription: response.subscription });
-      return response;
-    },
-    messagePresignFiles: async (atts: Att[], authMethod?: FcAuthMethods): Promise<R.FcMsgPresignFiles> => {
-      let response: R.FcMsgPresignFiles;
-      const lengths = atts.map(a => a.length);
-      if (!authMethod) {
-        response = await Api.internal.apiFcCall('message/presign_files', {
-          lengths,
-        }) as R.FcMsgPresignFiles;
-      } else if (authMethod === 'uuid') {
-        const authInfo = await Store.authInfo();
-        response = await Api.internal.apiFcCall('message/presign_files', {
-          account: authInfo.acctEmail,
-          uuid: authInfo.uuid,
-          lengths,
-        }) as R.FcMsgPresignFiles;
-      } else {
-        response = await Api.internal.apiFcCall('message/presign_files', {
-          message_token_account: authMethod.account,
-          message_token: authMethod.token,
-          lengths,
-        }) as R.FcMsgPresignFiles;
-      }
-      if (response.approvals && response.approvals.length === atts.length) {
-        return response;
-      }
-      throw new Error('Could not verify that all files were uploaded properly, please try again.');
-    },
-    messageConfirmFiles: (identifiers: string[]): Promise<R.FcMsgConfirmFiles> => Api.internal.apiFcCall('message/confirm_files', {
-      identifiers,
-    }),
-    messageUpload: async (encryptedDataArmored: string, authMethod?: FcAuthMethods): Promise<R.FcMsgUpload> => { // todo - DEPRECATE THIS. Send as JSON to message/store
-      if (encryptedDataArmored.length > 100000) {
-        throw new Error('Message text should not be more than 100 KB. You can send very long texts as attachments.');
-      }
-      const content = new Att({ name: 'cryptup_encrypted_message.asc', type: 'text/plain', data: Buf.fromUtfStr(encryptedDataArmored) });
-      if (!authMethod) {
-        return await Api.internal.apiFcCall('message/upload', { content }, 'FORM') as R.FcMsgUpload;
-      } else {
-        const authInfo = await Store.authInfo();
-        return await Api.internal.apiFcCall('message/upload', { account: authInfo.acctEmail, uuid: authInfo.uuid, content }, 'FORM') as R.FcMsgUpload;
-      }
-    },
-    messageToken: async (): Promise<R.FcMsgToken> => {
-      const authInfo = await Store.authInfo();
-      return await Api.internal.apiFcCall('message/token', { account: authInfo.acctEmail, uuid: authInfo.uuid }) as R.FcMsgToken;
-    },
-    messageExpiration: async (adminCodes: string[], addDays?: number): Promise<R.ApirFcMsgExpiration> => {
-      const authInfo = await Store.authInfo();
-      return await Api.internal.apiFcCall('message/expiration', {
-        account: authInfo.acctEmail,
-        uuid: authInfo.uuid,
-        admin_codes: adminCodes,
-        add_days: addDays || null, // tslint:disable-line:no-null-keyword
-      }) as R.ApirFcMsgExpiration;
-    },
-    messageReply: (short: string, token: string, from: string, to: string, subject: string, message: string): Promise<unknown> => Api.internal.apiFcCall('message/reply', {
-      short,
-      token,
-      from,
-      to,
-      subject,
-      message,
-    }),
-    messageContact: (sender: string, message: string, messageToken: FcAuthToken): Promise<unknown> => Api.internal.apiFcCall('message/contact', {
-      message_token_account: messageToken.account,
-      message_token: messageToken.token,
-      sender,
-      message,
-    }),
-    linkMessage: (short: string): Promise<R.FcLinkMsg> => Api.internal.apiFcCall('link/message', {
-      short,
-    }),
-    linkMe: (alias: string): Promise<R.FcLinkMe> => Api.internal.apiFcCall('link/me', {
-      alias,
-    }),
-  };
-
-  public static aws = {
-    s3Upload: (items: AwsS3UploadItem[], progressCb: ProgressCb) => {
-      const progress = Value.arr.zeroes(items.length);
-      const promises: Promise<void>[] = [];
-      if (!items.length) {
-        return Promise.resolve(promises);
-      }
-      for (const i of items.keys()) {
-        const fields = items[i].fields;
-        fields.file = new Att({ name: 'encrpted_attachment', type: 'application/octet-stream', data: items[i].att.getData() });
-        promises.push(Api.internal.apiCall(items[i].baseUrl, '', fields, 'FORM', {
-          upload: (singleFileProgress: number) => {
-            progress[i] = singleFileProgress;
-            Ui.event.prevent('spree', () => progressCb(Value.arr.average(progress)))();
-          }
-        }));
-      }
-      return Promise.all(promises);
-    },
-  };
-
   public static download = (url: string, progress?: ProgressCb): Promise<Buf> => new Promise((resolve, reject) => {
     const request = new XMLHttpRequest();
     request.open('GET', url, true);
@@ -627,7 +328,7 @@ export class Api {
       if (e instanceof Error) {
         throw e;
       }
-      if (Api.internal.isRawAjaxError(e)) {
+      if (Api.isRawAjaxError(e)) {
         throw AjaxError.fromXhr(e, req, stack);
       }
       throw new Error(`Unknown Ajax error (${String(e)}) type when calling ${req.url}`);
@@ -657,83 +358,58 @@ export class Api {
     };
   }
 
-  public static encodeAsMultipartRelated = (parts: Dict<string>) => { // todo - this could probably be achieved with emailjs-mime-builder
-    const boundary = 'this_sucks_' + Str.sloppyRandom(10);
-    let body = '';
-    for (const type of Object.keys(parts)) {
-      body += '--' + boundary + '\n';
-      body += 'Content-Type: ' + type + '\n';
-      if (type.includes('json')) {
-        body += '\n' + parts[type] + '\n\n';
-      } else {
-        body += 'Content-Transfer-Encoding: base64\n';
-        body += '\n' + btoa(parts[type]) + '\n\n';
-      }
-    }
-    body += '--' + boundary + '--';
-    return { contentType: 'multipart/related; boundary=' + boundary, body };
+  private static isRawAjaxError = (e: any): e is RawAjaxError => {
+    return e && typeof e === 'object' && typeof (e as RawAjaxError).readyState === 'number';
   }
 
-  private static internal = {
-    isStandardError: (e: any): e is StandardError => {
-      return e && typeof e === 'object' && (e as StandardError).hasOwnProperty('internal') && Boolean((e as StandardError).message);
-    },
-    isRawAjaxError: (e: any): e is RawAjaxError => {
-      return e && typeof e === 'object' && typeof (e as RawAjaxError).readyState === 'number';
-    },
-    apiCall: async (url: string, path: string, fields?: Dict<any>, fmt?: ReqFmt, progress?: ProgressCbs, headers?: Dict<string>, resFmt: ResFmt = 'json', method: ReqMethod = 'POST') => {
-      progress = progress || {} as ProgressCbs;
-      let formattedData: FormData | string | undefined;
-      let contentType: string | false;
-      if (fmt === 'JSON' && fields) {
-        formattedData = JSON.stringify(fields);
-        contentType = 'application/json; charset=UTF-8';
-      } else if (fmt === 'FORM' && fields) {
-        formattedData = new FormData();
-        for (const formFieldName of Object.keys(fields)) {
-          const a: Att | string = fields[formFieldName]; // tslint:disable-line:no-unsafe-any
-          if (a instanceof Att) {
-            formattedData.append(formFieldName, new Blob([a.getData()], { type: a.type }), a.name); // xss-none
-          } else {
-            formattedData.append(formFieldName, a); // xss-none
-          }
-        }
-        contentType = false;
-      } else if (!fmt && !fields && method === 'GET') {
-        formattedData = undefined;
-        contentType = false;
-      } else {
-        throw new Error('unknown format:' + String(fmt));
-      }
-      const req: JQueryAjaxSettings = {
-        xhr: Api.getAjaxProgressXhrFactory(progress),
-        url: url + path,
-        method,
-        data: formattedData,
-        dataType: resFmt,
-        crossDomain: true,
-        headers,
-        processData: false,
-        contentType,
-        async: true,
-        timeout: typeof progress!.upload === 'function' || typeof progress!.download === 'function' ? undefined : 20000, // substituted with {} above
-      };
-      const res = await Api.ajax(req, Catch.stackTrace());
-      if (res && typeof res === 'object' && typeof (res as StandardErrorRes).error === 'object' && (res as StandardErrorRes).error.message) {
-        throw new ApiErrorResponse(res as StandardErrorRes, req);
-      }
-      return res;
-    },
-    apiAttesterCall: (path: string, values?: Dict<any>, method: ReqMethod = 'POST'): Promise<any> => {
-      return Api.internal.apiCall('https://flowcrypt.com/attester/', path, values, 'JSON', undefined, { 'api-version': '3' }, 'json', method);
-    },
-    apiFcCall: (path: string, vals: Dict<any>, fmt: ReqFmt = 'JSON'): Promise<any> => {
-      return Api.internal.apiCall(Api.fc.url('api'), path, vals, fmt, undefined, { 'api-version': '3' });
-    },
-  };
+  private static isStandardError = (e: any): e is StandardError => {
+    return e && typeof e === 'object' && (e as StandardError).hasOwnProperty('internal') && Boolean((e as StandardError).message);
+  }
 
-  public static retreiveBlogPosts = async (): Promise<R.FcBlogPost[]> => {
-    return Api.ajax({ url: 'https://flowcrypt.com/feed', dataType: 'json' }, Catch.stackTrace());
+  protected static apiCall = async (
+    url: string, path: string, fields?: Dict<any>, fmt?: ReqFmt, progress?: ProgressCbs, headers?: Dict<string>, resFmt: ResFmt = 'json', method: ReqMethod = 'POST'
+  ) => {
+    progress = progress || {} as ProgressCbs;
+    let formattedData: FormData | string | undefined;
+    let contentType: string | false;
+    if (fmt === 'JSON' && fields) {
+      formattedData = JSON.stringify(fields);
+      contentType = 'application/json; charset=UTF-8';
+    } else if (fmt === 'FORM' && fields) {
+      formattedData = new FormData();
+      for (const formFieldName of Object.keys(fields)) {
+        const a: Att | string = fields[formFieldName]; // tslint:disable-line:no-unsafe-any
+        if (a instanceof Att) {
+          formattedData.append(formFieldName, new Blob([a.getData()], { type: a.type }), a.name); // xss-none
+        } else {
+          formattedData.append(formFieldName, a); // xss-none
+        }
+      }
+      contentType = false;
+    } else if (!fmt && !fields && method === 'GET') {
+      formattedData = undefined;
+      contentType = false;
+    } else {
+      throw new Error('unknown format:' + String(fmt));
+    }
+    const req: JQueryAjaxSettings = {
+      xhr: Api.getAjaxProgressXhrFactory(progress),
+      url: url + path,
+      method,
+      data: formattedData,
+      dataType: resFmt,
+      crossDomain: true,
+      headers,
+      processData: false,
+      contentType,
+      async: true,
+      timeout: typeof progress!.upload === 'function' || typeof progress!.download === 'function' ? undefined : 20000, // substituted with {} above
+    };
+    const res = await Api.ajax(req, Catch.stackTrace());
+    if (res && typeof res === 'object' && typeof (res as StandardErrorRes).error === 'object' && (res as StandardErrorRes).error.message) {
+      throw new ApiErrorResponse(res as StandardErrorRes, req);
+    }
+    return res;
   }
 
 }
