+++ extension/js/common/composer.ts
@@ -9,11 +9,13 @@ import { Value, Str } from './core/common.js';
 import { Att } from './core/att.js';
 import { BrowserMsg, Extension, BrowserWidnow } from './extension.js';
 import { Pgp, Pwd, FormatError, Contact, KeyInfo, PgpMsg } from './core/pgp.js';
-import { Api, R, ProgressCb, ProviderContactsQuery, PubkeySearchResult, SendableMsg, AwsS3UploadItem, ChunkedCb, AjaxError } from './api/api.js';
+import { Api, ProgressCb, ProviderContactsQuery, SendableMsg, ChunkedCb, AjaxError } from './api/api.js';
 import { Ui, Xss, AttUI, BrowserEventErrHandler, Env, AttLimits } from './browser.js';
 import { Mime, SendableMsgBody } from './core/mime.js';
-import { GoogleAuth } from './api/google.js';
+import { GoogleAuth, GmailRes } from './api/google.js';
 import { Buf } from './core/buf.js';
+import { PubkeySearchResult, Attester } from './api/attester.js';
+import { Backend, AwsS3UploadItem, BackendRes } from './api/backend.js';
 
 declare const openpgp: typeof OpenPGP;
 
@@ -35,11 +37,11 @@ interface ComposerAppFunctionsInterface {
   storageContactSave: (contact: Contact) => Promise<void>;
   storageContactSearch: (query: ProviderContactsQuery) => Promise<Contact[]>;
   storageContactObj: (email: string, name?: string, client?: string, pubkey?: string, pendingLookup?: boolean, lastUse?: number) => Promise<Contact>;
-  emailProviderDraftGet: (draftId: string) => Promise<R.GmailDraftGet | undefined>;
-  emailProviderDraftCreate: (acctEmail: string, mimeMsg: string, threadId?: string) => Promise<R.GmailDraftCreate>;
-  emailProviderDraftUpdate: (draftId: string, mimeMsg: string) => Promise<R.GmailDraftUpdate>;
-  emailProviderDraftDelete: (draftId: string) => Promise<R.GmailDraftDelete>;
-  emailProviderMsgSend: (msg: SendableMsg, renderUploadProgress: ProgressCb) => Promise<R.GmailMsgSend>;
+  emailProviderDraftGet: (draftId: string) => Promise<GmailRes.GmailDraftGet | undefined>;
+  emailProviderDraftCreate: (acctEmail: string, mimeMsg: string, threadId?: string) => Promise<GmailRes.GmailDraftCreate>;
+  emailProviderDraftUpdate: (draftId: string, mimeMsg: string) => Promise<GmailRes.GmailDraftUpdate>;
+  emailProviderDraftDelete: (draftId: string) => Promise<GmailRes.GmailDraftDelete>;
+  emailProviderMsgSend: (msg: SendableMsg, renderUploadProgress: ProgressCb) => Promise<GmailRes.GmailMsgSend>;
   emailEroviderSearchContacts: (query: string, knownContacts: Contact[], multiCb: ChunkedCb) => void;
   emailProviderDetermineReplyMsgHeaderVariables: () => Promise<undefined | { lastMsgId: string, headers: { 'In-Reply-To': string, 'References': string } }>;
   emailProviderExtractArmoredBlock: (msgId: string) => Promise<string>;
@@ -726,13 +728,13 @@ export class Composer {
   }
 
   private uploadAttsToFc = async (atts: Att[], subscription: Subscription): Promise<string[]> => {
-    const pfRes: R.FcMsgPresignFiles = await Api.fc.messagePresignFiles(atts, subscription.active ? 'uuid' : undefined);
+    const pfRes: BackendRes.FcMsgPresignFiles = await Backend.fc.messagePresignFiles(atts, subscription.active ? 'uuid' : undefined);
     const items: AwsS3UploadItem[] = [];
     for (const i of pfRes.approvals.keys()) {
       items.push({ baseUrl: pfRes.approvals[i].base_url, fields: pfRes.approvals[i].fields, att: atts[i] });
     }
-    await Api.aws.s3Upload(items, this.renderUploadProgress);
-    const { admin_codes, confirmed } = await Api.fc.messageConfirmFiles(items.map(item => item.fields.key));
+    await Backend.aws.s3Upload(items, this.renderUploadProgress);
+    const { admin_codes, confirmed } = await Backend.fc.messageConfirmFiles(items.map(item => item.fields.key));
     if (!confirmed || confirmed.length !== items.length) {
       throw new Error('Attachments did not upload properly, please try again');
     }
@@ -768,7 +770,7 @@ export class Composer {
     }
     let response;
     try {
-      response = await Api.fc.messageToken();
+      response = await Backend.fc.messageToken();
     } catch (msgTokenErr) {
       if (Api.err.isAuthErr(msgTokenErr)) {
         if (await Ui.modal.confirm('Your FlowCrypt account information is outdated, please review your account settings.')) {
@@ -830,7 +832,7 @@ export class Composer {
     if (pwd) {
       // this is used when sending encrypted messages to people without encryption plugin, the encrypted data goes through FlowCrypt and recipients get a link
       // admin_code stays locally and helps the sender extend life of the message or delete it
-      const { short, admin_code } = await Api.fc.messageUpload(encryptedBody['text/plain']!, subs.active ? 'uuid' : undefined);
+      const { short, admin_code } = await Backend.fc.messageUpload(encryptedBody['text/plain']!, subs.active ? 'uuid' : undefined);
       const storage = await Store.getAcct(this.urlParams.acctEmail, ['outgoing_language']);
       encryptedBody = this.fmtPwdProtectedEmail(short, encryptedBody, pubkeys, atts, storage.outgoing_language || 'EN');
       encryptedBody = this.formatEmailTextFooter(encryptedBody);
@@ -852,7 +854,7 @@ export class Composer {
     if (this.S.cached('icon_pubkey').is('.active')) {
       msg.atts.push(Att.keyinfoAsPubkeyAtt(await this.app.storageGetKey(this.urlParams.acctEmail)));
     }
-    let msgSentRes: R.GmailMsgSend;
+    let msgSentRes: GmailRes.GmailMsgSend;
     try {
       msgSentRes = await this.app.emailProviderMsgSend(msg, this.renderUploadProgress);
     } catch (e) {
@@ -883,7 +885,7 @@ export class Composer {
     } else {
       try {
         this.debug(`lookupPubkeyFromDbOrKeyserverAndUpdateDbIfneeded.1`);
-        const lookupResult = await Api.attester.lookupEmail(email);
+        const lookupResult = await Attester.attester.lookupEmail(email);
         this.debug(`lookupPubkeyFromDbOrKeyserverAndUpdateDbIfneeded.2`);
         if (lookupResult && email) {
           if (lookupResult.pubkey) {
