+++ extension/js/common/settings.ts
@@ -12,6 +12,7 @@ import { Rules } from './rules.js';
 import { Api } from './api/api.js';
 import { Pgp } from './core/pgp.js';
 import { Google, GoogleAuth } from './api/google.js';
+import { Attester } from './api/attester.js';
 
 declare const openpgp: typeof OpenPGP;
 declare const zxcvbn: Function; // tslint:disable-line:ban-types
@@ -53,10 +54,10 @@ export class Settings {
   }
 
   static submitPubkeys = async (acctEmail: string, addresses: string[], pubkey: string) => {
-    await Api.attester.initialLegacySubmit(acctEmail, pubkey);
+    await Attester.attester.initialLegacySubmit(acctEmail, pubkey);
     const aliases = addresses.filter(a => a !== acctEmail);
     if (aliases.length) {
-      await Promise.all(aliases.map(a => Api.attester.initialLegacySubmit(a, pubkey)));
+      await Promise.all(aliases.map(a => Attester.attester.initialLegacySubmit(a, pubkey)));
     }
   }
